Sua loi kieu du lieu long trong cac repository, cau hinh cors cho fe

+++ src/main/java/com/example/webbanghang/repository/ProductAttributeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 
 import com.example.webbanghang.model.entity.ProductAttribute;
 @Repository
-public interface ProductAttributeRepository extends JpaRepository<ProductAttribute, Long> {
+public interface ProductAttributeRepository extends JpaRepository<ProductAttribute, Integer> {
 
 }
